# services/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/prompts": {
      "post": {
        "operationId": "PromptLibraryController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Prompt created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "409": {
            "description": "Prompt slug already exists"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new prompt",
        "tags": [
          "Prompt Library"
        ]
      },
      "get": {
        "operationId": "PromptLibraryController_list",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List prompts",
        "tags": [
          "Prompt Library"
        ]
      }
    },
    "/api/v1/prompts/categories": {
      "get": {
        "operationId": "PromptLibraryController_getCategories",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List prompt categories",
        "tags": [
          "Prompt Library"
        ]
      }
    },
    "/api/v1/prompts/{id}": {
      "get": {
        "operationId": "PromptLibraryController_getById",
        "parameters": [],
        "responses": {
          "404": {
            "description": "Prompt not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get prompt by ID",
        "tags": [
          "Prompt Library"
        ]
      },
      "patch": {
        "operationId": "PromptLibraryController_update",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update prompt metadata",
        "tags": [
          "Prompt Library"
        ]
      },
      "delete": {
        "operationId": "PromptLibraryController_delete",
        "parameters": [],
        "responses": {
          "204": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft-delete a prompt",
        "tags": [
          "Prompt Library"
        ]
      }
    },
    "/api/v1/prompts/{id}/versions": {
      "get": {
        "operationId": "PromptLibraryController_getVersions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get prompt version history",
        "tags": [
          "Prompt Library"
        ]
      },
      "post": {
        "operationId": "PromptLibraryController_createVersion",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new version of a prompt",
        "tags": [
          "Prompt Library"
        ]
      }
    },
    "/api/v1/prompts/{id}/validate": {
      "get": {
        "operationId": "PromptLibraryController_validate",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Validate prompt templates",
        "tags": [
          "Prompt Library"
        ]
      }
    },
    "/api/v1/prompts/{id}/render": {
      "post": {
        "operationId": "PromptLibraryController_render",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Render a prompt with variables",
        "tags": [
          "Prompt Library"
        ]
      }
    },
    "/api/v1/prompts/preview": {
      "post": {
        "operationId": "PromptLibraryController_preview",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Preview template rendering",
        "tags": [
          "Prompt Library"
        ]
      }
    },
    "/api/v1/prompts/{id}/publish": {
      "post": {
        "operationId": "PromptLibraryController_publish",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Publish a prompt",
        "tags": [
          "Prompt Library"
        ]
      }
    },
    "/api/v1/prompts/{id}/archive": {
      "post": {
        "operationId": "PromptLibraryController_archive",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Archive a prompt",
        "tags": [
          "Prompt Library"
        ]
      }
    },
    "/api/v1/prompts/{id}/restore": {
      "post": {
        "operationId": "PromptLibraryController_restore",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Restore an archived prompt",
        "tags": [
          "Prompt Library"
        ]
      }
    },
    "/api/v1/prompts/{id}/rollback": {
      "post": {
        "operationId": "PromptLibraryController_rollback",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Rollback to a previous version",
        "tags": [
          "Prompt Library"
        ]
      }
    },
    "/api/v1/prompts/{id}/compare": {
      "get": {
        "operationId": "PromptLibraryController_compare",
        "parameters": [
          {
            "name": "versionB",
            "required": true,
            "in": "query",
            "schema": {}
          },
          {
            "name": "versionA",
            "required": true,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Compare two versions of a prompt",
        "tags": [
          "Prompt Library"
        ]
      }
    },
...
  },
  "info": {
    "title": "Atlas AI API",
    "description": "Enterprise AI Platform API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}